tims-main/pyproject.toml
[project]
name = "tims"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bitstring>=4.3.1",
    "cocotb>=2.0.0",
    "cocotbext-spi>=0.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pyside6>=6.10.1",
    "pytest>=9.0.1",
    "scipy>=1.16.3",
]
